refactor: centralize config module imports to use direct import pattern
