+++ gimpopenvino/plugins/openvino_utils/tools/stable_diffusion_ov_server.py
@@ -20,13 +20,12 @@ import numpy as np
 import psutil
 import threading
 
-sys.path.extend([os.path.join(os.path.dirname(os.path.realpath(__file__)), "..", "..", "..")])
-from gimpopenvino import config
-
 sys.path.extend([os.path.join(os.path.dirname(os.path.realpath(__file__)), "openvino_common")])
 sys.path.extend([os.path.join(os.path.dirname(os.path.realpath(__file__)), "..","tools")])
+sys.path.extend([os.path.join(os.path.dirname(os.path.realpath(__file__)), "..")])
 
 from gimpopenvino.plugins.openvino_utils.tools.tools_utils import get_weight_path, SDOptionCache,config_path_dir
+import config
 
 from diffusers.schedulers import DDIMScheduler, LMSDiscreteScheduler, LCMScheduler, EulerDiscreteScheduler
 from models_ov.stable_diffusion_engine import StableDiffusionEngineAdvanced, StableDiffusionEngine, LatentConsistencyEngine, StableDiffusionEngineReferenceOnly
