+++ gimpopenvino/plugins/stable_diffusion_ov/stable_diffusion_ov.py
@@ -25,13 +25,11 @@ from enum import IntEnum
 import glob
 from pathlib import Path
 
-sys.path.extend([os.path.join(os.path.dirname(os.path.realpath(__file__)), "..", "..")])
-from gimpopenvino import config
-
 sys.path.extend([os.path.join(os.path.dirname(os.path.realpath(__file__)), "..","openvino_utils")])
 from plugin_utils import *
 from model_management_window import ModelManagementWindow
 from tools.tools_utils import base_model_dir, config_path_dir, SDOptionCache
+import config
 
 _ = gettext.gettext
 image_paths = {
